+++ src/controllers/lab/lab_controllers.js
@@ -1,148 +1,3 @@
-
-
-// O========================================================================================O
-
-/*
-    O=================================================================O
-    |   Funções de control relacionadas a edição de laboratórios      |
-    O=================================================================O
- 
-    Funções relacionadas a edição de laboratórios:	
-    - [X] EditLabName;
-    - [X] EditLabCapacity;
-*/
-
-// O========================================================================================O
-
-// Função para editar o nome do laboratório:
-const EditLabName = async (req, res) => {
-  const { lab_id, sala } = req.body;
-
-  // Verificando se o usuário existe:
-  const token = req.headers["x-access-token"];
-  const userID = jwt.decode(token).userID;
-
-  const GetUserByID = await userModels.getUserByID(userID);
-
-  if (GetUserByID.status === false) {
-    return res.status(400).json({
-      status: false,
-      message: "Usuário não encontrado!",
-    });
-  }
-
-  // Verificando se o laboratório existe:
-  const GetLabByID = await labModels.GetLabById(lab_id);
-
-  if (GetLabByID.status === false) {
-    return res.status(400).json({
-      status: false,
-      message: "Laboratório não encontrado!",
-    });
-  }
-
-  // Verificando permissão do usuário:
-  const GetLabUser = await labModels.GetLabUser(lab_id, userID);
-
-  if (
-    GetLabUser.status === false ||
-    GetLabUser.data.AdminLevel === 1 ||
-    GetUserByID.userData.ID_campus !== GetLabByID.data.ID_campus
-  ) {
-    return res.status(400).json({
-      status: false,
-      message: "Usuário não tem permissão para editar o laboratório!",
-    });
-  }
-
-  // Verificando se o novo nome já existe:
-  const GetLabByName = await labModels.GetLabByName(
-    sala,
-    GetLabByID.data.ID_campus
-  );
-
-  if (GetLabByName.status === true) {
-    return res.status(400).json({
-      status: false,
-      message: "Nome de laboratório já existe!",
-    });
-  }
-
-  // Editando o nome do laboratório:
-  const EditLabName = await labModels.EditLabName(lab_id, sala);
-
-  if (EditLabName.status === false) {
-    return res.status(400).json({
-      status: false,
-      message: "Erro ao editar nome do laboratório!",
-    });
-  }
-
-  return res.status(200).json({
-    status: true,
-    message: "Nome do laboratório editado com sucesso!",
-  });
-};
-
-// +------------------------------------------------------------------------------------------+
-
-// Função para editar a capacidade do laboratório:
-const EditLabCapacity = async (req, res) => {
-  const { lab_id, capacity } = req.body;
-
-  // Verificando se o usuário existe:
-  const token = req.headers["x-access-token"];
-  const userID = jwt.decode(token).userID;
-
-  const GetUserByID = await userModels.getUserByID(userID);
-
-  if (GetUserByID.status === false) {
-    return res.status(400).json({
-      status: false,
-      message: "Usuário não encontrado!",
-    });
-  }
-
-  // Verificando se o laboratório existe:
-  const GetLabByID = await labModels.GetLabById(lab_id);
-
-  if (GetLabByID.status === false) {
-    return res.status(400).json({
-      status: false,
-      message: "Laboratório não encontrado!",
-    });
-  }
-
-  // Verificando permissão do usuário:
-  const GetLabUser = await labModels.GetLabUser(lab_id, userID);
-
-  if (
-    GetLabUser.status === false ||
-    GetLabUser.data.AdminLevel === 1 ||
-    GetUserByID.userData.ID_campus !== GetLabByID.data.ID_campus
-  ) {
-    return res.status(400).json({
-      status: false,
-      message: "Usuário não tem permissão para editar o laboratório!",
-    });
-  }
-
-  // Editando a capacidade do laboratório:
-  const EditLabCapacity = await labModels.EditLabCapacity(lab_id, capacity);
-
-  if (EditLabCapacity.status === false) {
-    return res.status(400).json({
-      status: false,
-      message: "Erro ao editar capacidade do laboratório!",
-    });
-  }
-
-  return res.status(200).json({
-    status: true,
-    message: "Capacidade do laboratório editada com sucesso!",
-  });
-};
-
 // O========================================================================================O
 
 /*
